Move IMethodSymbol
